+++ documentation/docs/release-notes.md
@@ -9,6 +9,7 @@ title: Release notes&#58;
 - Upgrade to Shiro v2.1 for security reasons. Defaults remain the ones of Shiro v1.13 though and will change in a future major version.
 - Add the `encryptionRequired` flag (by default `false`) for the `JWtAuthenticator`
 - `KeystoreGenerator` and its implem `FileSystemKeystoreGenerator` are now in the core
+- OIDC protocol supports the `client_secret_jwt` authentication method
 - Support of the OpenID Connect Federation protocol v1.0
 
 **v6.3.3**: